Made the directory structure more analogous to BloomReader-RN
