+++ src/storage/BookStorage.ts
@@ -1,5 +1,6 @@
 import * as FileSystem from "expo-file-system";
 import { unzip } from "react-native-zip-archive";
+import { safeUnlink } from "../util/FileUtil";
 
 export const OPEN_BOOK_DIR = FileSystem.cacheDirectory + "openBook";
 
@@ -11,27 +12,4 @@ export async function openBookForReading(
 
     //console.log(`starting unzip from ${bookFilePath}\nto ${OPEN_BOOK_DIR}`);
     return unzip(bookFilePath, OPEN_BOOK_DIR);
-}
-
-/**
- * Deletes a directory or file
- * @remarks Based on rnfsSafeUnlink in BloomReader-RN
- */
-export async function safeUnlink(path: string): Promise<void> {
-    try {
-        await FileSystem.deleteAsync(path, {
-            idempotent: true,
-        });
-
-        const fileInfo = await FileSystem.getInfoAsync(path);
-        if (fileInfo.exists) {
-            throw new Error(`Tried to delete ${path}, but it's still there!`);
-        }
-    } catch (err) {
-        console.error({
-            logMessage: `[safeUnlink] Error deleting file: ${path}\n${JSON.stringify(
-                err
-            )}`,
-        });
-    }
 }
